bug fix in config (extra semi colon for sublister) updated install to not install photon

+++ setup/install.sh
@@ -36,17 +36,18 @@ if [ ! -f /opt/Sublist3r/sublist3r.py ]; then
     pip install -r requirements.txt
 fi
 
-echo "[+] Downloading Photon Web Spider to /opt/Photon/photon.py"
-cd /opt/
-git clone https://github.com/s0md3v/Photon.git
-cd Photon
-pip install -r requirements.txt
+if [ ! -f /opt/Photon/photon.py ]; then
+    echo "[+] Downloading Photon Web Spider to /opt/Photon/photon.py"
+    cd /opt/
+    git clone https://github.com/s0md3v/Photon.git
+    cd Photon
+    pip install -r requirements.txt
+fi
 
 cd $CELERYSTALK_DIR
 cp bash_completion_file /etc/bash_completion.d/celerystalk.sh
-exec bash
 ../celerystalk -h
 echo ""
 echo "[+] Back up a directory and you are ready to go."
-
+exec bash
